python-service/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database and Async Support
greenlet>=2.0.0

# AI and ML
openai==1.10.0
langchain==0.1.4
tiktoken==0.5.2
numpy==1.26.3
scikit-learn==1.4.0

# Excel Processing
openpyxl==3.1.2
pandas==2.1.4
xlrd==2.0.1      # For reading .xls files
xlwt==1.3.0      # For writing .xls files
oletools==0.60.2
python-magic==0.4.27
formulas==1.2.6  # Excel 수식 실시간 계산 엔진

# AI Model Providers for Failover System
anthropic>=0.18.0  # Claude API
groq>=0.4.0  # Groq high-speed inference
google-cloud-aiplatform>=1.106.0  # Google Vertex AI for embeddings

# Image Processing
Pillow==10.2.0
pytesseract==0.3.10
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80

# Language Detection for OCR
langdetect==1.0.9

# OpenAI already included above (openai==1.10.0)

# Database
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1

# Redis
redis==5.0.1

# Vector Database
pgvector==0.2.4

# Environment and Config
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1
aiofiles==23.2.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
Jinja2==3.1.3

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Monitoring
prometheus-client==0.19.0
psutil==5.9.8

# Forum data collection
praw==7.7.1
selenium==4.15.2
python-dotenv==1.0.0
structlog==24.1.0
PyJWT==2.8.0
